Reorganize file structure (#263)
Reorganizes the file structure to be organized by concept
