+++ src/authenticated/project/GeneralProjectPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MarkdownFileRenderer } from '../utils';
+import MarkdownFileRenderer from '../../shared-components/MarkdownFileRenderer';
 
 const GeneralProjectPage = () => (
   <MarkdownFileRenderer fileName="general_project_info" />
