+++ src/shared-components/ErrorBoundary.js
