+++ src/utils/index.js
@@ -1,7 +1,3 @@
 export { default as withLoggedInUser } from './withLoggedInUser';
 export { default as withAuthSession } from './withAuthSession';
 export { default as getEnvironmentVariables } from './getEnvironmentVariables';
-export { default as MarkdownFileRenderer } from './MarkdownFileRenderer';
-export {
-  default as GoogleAnalyticsPageTracker
-} from './GoogleAnalyticsPageTracker';
